move : relocated station DTOs
